Refactor listeners table header to separate component
