+++ ui/e2e/events/listeners.spec.ts
@@ -109,6 +109,8 @@ test("it paginates event listeners", async ({ page }) => {
     `/n/${namespace}/events/listeners`
   );
 
+  debugger;
+
   await expect(
     page.getByRole("cell", { name: "start workflow" }),
     "it renders the expected number of items on page 1"
