+++ ui/src/pages/namespace/Events/Listeners/ListenersList.tsx
@@ -1,17 +1,9 @@
 import { Dispatch, SetStateAction } from "react";
-import {
-  NoPermissions,
-  NoResult,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeaderCell,
-  TableRow,
-} from "~/design/Table";
+import { NoPermissions, NoResult, TableCell, TableRow } from "~/design/Table";
 
 import { Antenna } from "lucide-react";
 import { Card } from "~/design/Card";
+import ListenersTable from "./Table";
 import { Pagination } from "~/components/Pagination";
 import Row from "./Row";
 import { itemsPerPage } from ".";
@@ -37,57 +29,36 @@ const ListenersList = ({
   return (
     <div className="flex grow flex-col gap-y-3 p-5">
       <Card>
-        <Table className="border-gray-5 dark:border-gray-dark-5">
-          <TableHead>
+        <ListenersTable>
+          {isAllowed ? (
+            <>
+              {noResults ? (
+                <TableRow className="hover:bg-inherit dark:hover:bg-inherit">
+                  <TableCell colSpan={6}>
+                    <NoResult icon={Antenna}>
+                      {t("pages.events.listeners.empty.noResults")}
+                    </NoResult>
+                  </TableCell>
+                </TableRow>
+              ) : (
+                data?.data.map((listener, i) => (
+                  <Row
+                    listener={listener}
+                    key={i}
+                    namespace={listener.namespace}
+                    data-testid={`listener-row-${i}`}
+                  />
+                ))
+              )}
+            </>
+          ) : (
             <TableRow className="hover:bg-inherit dark:hover:bg-inherit">
-              <TableHeaderCell>
-                {t("pages.events.listeners.tableHeader.type")}
-              </TableHeaderCell>
-              <TableHeaderCell>
-                {t("pages.events.listeners.tableHeader.target")}
-              </TableHeaderCell>
-              <TableHeaderCell>
-                {t("pages.events.listeners.tableHeader.mode")}
-              </TableHeaderCell>
-              <TableHeaderCell>
-                {t("pages.events.listeners.tableHeader.createdAt")}
-              </TableHeaderCell>
-              <TableHeaderCell>
-                {t("pages.events.listeners.tableHeader.eventTypes")}
-              </TableHeaderCell>
+              <TableCell colSpan={6}>
+                <NoPermissions>{noPermissionMessage}</NoPermissions>
+              </TableCell>
             </TableRow>
-          </TableHead>
-          <TableBody>
-            {isAllowed ? (
-              <>
-                {noResults ? (
-                  <TableRow className="hover:bg-inherit dark:hover:bg-inherit">
-                    <TableCell colSpan={6}>
-                      <NoResult icon={Antenna}>
-                        {t("pages.events.listeners.empty.noResults")}
-                      </NoResult>
-                    </TableCell>
-                  </TableRow>
-                ) : (
-                  data?.data.map((listener, i) => (
-                    <Row
-                      listener={listener}
-                      key={i}
-                      namespace={listener.namespace}
-                      data-testid={`listener-row-${i}`}
-                    />
-                  ))
-                )}
-              </>
-            ) : (
-              <TableRow className="hover:bg-inherit dark:hover:bg-inherit">
-                <TableCell colSpan={6}>
-                  <NoPermissions>{noPermissionMessage}</NoPermissions>
-                </TableCell>
-              </TableRow>
-            )}
-          </TableBody>
-        </Table>
+          )}
+        </ListenersTable>
       </Card>
       {showPagination && (
         <Pagination
